+++ app/assets/javascripts/Components/markus_with_selection_hoc.jsx
@@ -19,6 +19,7 @@ function withSelection(WrappedComponent) {
       this.state = {
         selection: [],
         selectAll: false,
+        last_selected: null
       }
     }
 
@@ -29,22 +30,50 @@ function withSelection(WrappedComponent) {
       if (isNaN(key)) {
         return;
       }
-      let selection = [
-        ...this.state.selection
-      ];
-      const keyIndex = selection.indexOf(key);
-      if (keyIndex >= 0) {
-        selection.splice(keyIndex, 1);
+      let selection = [...this.state.selection];
+      let last_selected = null;
+      if (shift && this.state.last_selected !== null && this.state.last_selected !== key) {
+        // we need to get at the internals of ReactTable
+        const wrappedInstance = this.wrapped.checkboxTable.getWrappedInstance();
+        // the 'sortedData' property contains the currently accessible records based on the filter and sort
+        const currentRecords = wrappedInstance.getResolvedState().sortedData;
+        // find all keys between the last key selected and the current key selected
+        const keys_in_range = new Set;
+        currentRecords.some((record) => {
+          if (record._id === this.state.last_selected || record._id === key) {
+            keys_in_range.add(record._id);
+            if (keys_in_range.size > 1) {
+              return true; // acts like a break statement
+            }
+          } else if (!!keys_in_range.size) {
+            keys_in_range.add(record._id);
+          }
+          return false;
+        })
+        if (selection.indexOf(this.state.last_selected ) < 0) {
+          // if the last key selected unchecked the checkbox: remove all keys in range from the selection array
+          selection = [...new Set(selection.filter(x => !keys_in_range.has(x)))]
+        } else {
+          // otherwise add all keys in range to the selection array
+          selection = [...new Set([...selection, ...keys_in_range])]
+        }
       } else {
-        selection.push(key);
+        const keyIndex = selection.indexOf(key);
+        if (keyIndex >= 0) {
+          selection.splice(keyIndex, 1);
+        } else {
+          selection.push(key);
+        }
+        last_selected = key;
       }
       // update the state
-      this.setState({ selection });
+      this.setState({ selection, last_selected });
     }
 
     toggleAll() {
       const selectAll = !this.state.selectAll;
       const selection = [];
+      const last_selected = null;
       if (selectAll) {
         // we need to get at the internals of ReactTable
         const wrappedInstance = this.wrapped.checkboxTable.getWrappedInstance();
@@ -55,7 +84,7 @@ function withSelection(WrappedComponent) {
           selection.push(item._original._id);
         })
       }
-      this.setState({ selectAll, selection });
+      this.setState({ selectAll, selection, last_selected });
     }
 
     isSelected(key) {
@@ -63,7 +92,7 @@ function withSelection(WrappedComponent) {
     }
 
     resetSelection() {
-      this.setState({selectAll: false, selection: []});
+      this.setState({selectAll: false, selection: [], last_selected: null});
     }
 
     getCheckboxProps() {